+++ src/css/header.css
@@ -57,7 +57,7 @@
   margin-top: 2px;
 }
 
-.header-burger-btn {
+.open-menu {
   padding: 0;
   border: none;
   background-color: transparent;
