+++ src/css/hero.css
@@ -62,7 +62,18 @@
   background-repeat: no-repeat;
   background-size: cover;
 }
-
+.hero-img-rec {
+  min-width: 335px;
+  min-height: 366px;
+  @media screen and (min-width: 768px) {
+    min-width: 704px;
+    min-height: 512px;
+  }
+  @media screen and (min-width: 1280px) {
+    min-width: 1150px;
+    min-height: 500px;
+  }
+}
 @media screen and (min-width: 768px) {
   .hero-container {
     flex-direction: row;
