+++ src/css/mobile-menu.css
@@ -59,7 +59,7 @@
   }
 }
 
-.menu-close-btn {
+.menu-close {
   align-self: flex-end;
   width: 34px;
   height: 34px;
