+++ src/css/styles.css
@@ -6,6 +6,7 @@
 */
 /* Common styles */
 @import url('./reset.css');
+@import url('./normalize.css');
 @import url('./fonts.css');
 @import url('./base.css');
 @import url('./container.css');
