+++ src/partials/header.html
@@ -91,7 +91,7 @@
         </a>
       </li>
     </ul>
-    <button type="button" class="header-burger-btn" onclick="showDiv(event)">
+    <button type="button" class="open-menu" onclick="showDiv(event)">
       <svg
         class="header-burger-icon"
         width="34"
