+++ src/partials/mobile-menu.html
@@ -1,5 +1,5 @@
 <div class="mobile-menu-container" id="mobile-menu">
-  <button type="button" class="menu-close-btn" onclick="hideDiv()">
+  <button type="button" class="menu-close" onclick="hideDiv()">
     <svg class="menu-close-icon" width="20" height="20" viewBox="0 0 20 20">
       <use href="img/sprite.svg#menu-close"></use>
     </svg>
